Only add geometry columns if using spatial extensions

+++ db/migrate/20100415183707_add_geometry_to_stop_areas.rb
@@ -1,11 +1,17 @@
 class AddGeometryToStopAreas < ActiveRecord::Migration
   def self.up
-    add_column :stop_areas, :coords, :point, :srid => 27700
-    add_index :stop_areas, :coords, :spatial => true
+    spatial_extensions = MySociety::Config.getbool('USE_SPATIAL_EXTENSIONS', false) 
+    if spatial_extensions
+      add_column :stop_areas, :coords, :point, :srid => 27700
+      add_index :stop_areas, :coords, :spatial => true
+    end
   end
 
   def self.down
-    remove_index :stop_areas, :coords
-    remove_column :stop_areas, :coords
+    spatial_extensions = MySociety::Config.getbool('USE_SPATIAL_EXTENSIONS', false) 
+    if spatial_extensions
+      remove_index :stop_areas, :coords
+      remove_column :stop_areas, :coords
+    end
   end
 end
